requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
